requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
langchain
langchain-google-genai